Added comments to code and renamed things
