--- packages/changed-elements-react/src/widgets/V2Widget/models/JobStatus.ts
@@ -1,11 +1,0 @@
-export type JobStatus = "Unknown" | "Available" | "Not Processed" | "Processing" | "Error" | "Queued"; //todo add queued status
-
-export type JobProgress = {
-  currentProgress: number;
-  maxProgress: number;
-};
-
-export type JobStatusAndJobProgress = {
-  jobStatus: JobStatus;
-  jobProgress: JobProgress;
-};
